refactor: 디렉토리 구조 변경
- 도메인에 저장소를 포함
- 컨트롤러와 dto의 상위 디렉토리를 web으로 설정
