+++ src/main/java/codesquad/springcafe/domain/article/ArticleRepository.java
@@ -1,6 +1,5 @@
-package codesquad.springcafe.repository;
+package codesquad.springcafe.domain.article;
 
-import codesquad.springcafe.domain.Article;
 import org.springframework.stereotype.Repository;
 
 import java.util.Collections;
