+++ src/main/java/codesquad/springcafe/web/dto/UserUpdateDto.java
@@ -1,4 +1,4 @@
-package codesquad.springcafe.dto;
+package codesquad.springcafe.web.dto;
 
 public class UserUpdateDto {
